% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphegwasData.R
\docType{data}
\name{bppasn}
\alias{bppasn}
\title{LD Block data summarized in Approximately independent linkage disequilibrium blocks in human populations}
\format{
A dataframe with simulated GWAS results (simulated using plink 1.9)
\itemize{
  \item chr is the chromosome number
  \item start is the start position of the chromosome
  \item stop is the end position of the chromosome
}
}
\usage{
data(bppasn)
}
\value{
A data frame
}
\description{
The dataset 'bppasn' is a data.frame object with the LD blocks for asian population.
}
\keyword{datasets}
