% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWAS2Data.R
\docType{data}
\name{bppeur}
\alias{bppeur}
\title{LD Block data summarized in Approximately independent linkage disequilibrium blocks in human populations}
\format{
A dataframe with simulated GWAS results (simulated using plink 1.9)
\itemize{
  \item chr is the chromosome number
  \item start is the start position of the chromosome
  \item stop is the end position of the chromosome
}
}
\usage{
data(bppeur)
}
\value{
A data frame
}
\description{
The dataset 'bppeur' is a data.frame object with the LD blocks for European population.
}
\keyword{datasets}
