% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWAS2.R
\name{fastprocessphegwas}
\alias{fastprocessphegwas}
\title{Prepare the dataframe to pass to landscape function}
\usage{
fastprocessphegwas(phenos, LDblock = FALSE, LDpop = "eur")
}
\arguments{
\item{phenos}{Vector of names of dataframes that need to do PheGWAS on. Arrange the dataframe in the order how the the phenotypes should align in y axis}

\item{LDblock}{If want to pass a custom LDblock file for division of BP groups (applicable only for chromosomal level)}
}
\value{
A processed dataframe to pass to PheGWAS landscape function
}
\description{
Prepare the dataframe to pass to landscape function
}
\details{
Make sure there are no duplicate rsid's in any of the dataframe, If there aremake sure to resolve it before passing it to this function.
}
\examples{
phenos <- c("HDL", "LDL", "TRIGS", "TOTALCHOLESTROL")
## y is ready to be passed to function landscape
y <- fastprocessphegwas(phenos)
}
\author{
George Gittu
}
