% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphegwasData.R
\docType{data}
\name{ibd}
\alias{ibd}
\title{IBD GWAS summary dataset}
\format{
A dataframe with simulated GWAS results (simulated using plink 1.9)
\itemize{
  \item CHR is the chromosome number
  \item BP is the position of the chromosome
  \item rsid The SNP ID associated with that position and chromsome
  \item A1
  \item A2
  \item beta is the p_value of the trait
  \item se
  \item P.value is the p_value of the trait
}
}
\usage{
data(ibd)
}
\value{
A data frame
}
\description{
The dataset 'ibd' is a data.frame object with the summary results of a GWAS analysis.
}
\keyword{datasets}
