% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWAS2.R
\name{iphegwas}
\alias{iphegwas}
\title{Prepare the dataframe to pass to landscape function}
\usage{
iphegwas(phenos, dentogram = FALSE)
}
\arguments{
\item{phenos}{Vector of names of dataframes that need to do iPheGWAS on.}

\item{dentogram}{to show structural differences}
}
\value{
A processed dataframe to pass to PheGWAS landscape function
}
\description{
Prepare the dataframe to pass to landscape function
}
\details{
Make sure there are no duplicate rsid's in any of the dataframe, If there aremake sure to resolve it before passing it to this function.
}
\examples{
phenos <- c("HDL", "LDL", "TRIGS", "TOTALCHOLESTROL")
iphegwas(phenos)
iphegwas(phenos,dentogram = TRUE)
}
\author{
George Gittu
}
