% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWAS2.R
\name{landscapefast}
\alias{landscapefast}
\title{Interactive 3-D association landscape for many phenotypes}
\usage{
landscapefast(
  d,
  sliceval = 7,
  chromosome = FALSE,
  pop = "GBR",
  R2 = 0.75,
  D = 0.75,
  calculateLD = FALSE,
  mutualLD = FALSE,
  phenos,
  geneview = FALSE,
  levelsdown = 0
)
}
\arguments{
\item{d}{DataFrame output from processphegwas}

\item{sliceval}{Integer to indicate value of -log10(p) to do the sectionalcut. Usually value > -log10 6 is considered to be significant}

\item{chromosome}{Integer to indicate the chromosome number thats interested, If not given entire chromosome is given}

\item{pop}{The population to select for calculation the LD (default GB)}

\item{R2}{The value to set to calculate LD}

\item{D}{THe value to set to calcualte LD}

\item{calculateLD}{This shoudld be set to true if the calcualte LD logic needed to be added to the plot}

\item{mutualLD}{Calcualte the mutual LD SNP between the phenotypes}

\item{phenos}{Vector of names of dataframes that need to do PheGWAS on. Arrange the dataframe in the order how the the phenotypes should align in y axis
chromosome view the max peak is selected}

\item{geneview}{This checks for the common genes across the section}

\item{levelsdown}{Used to find the independant signals}

\item{LDblock}{If want to pass a custom LDblock file for division of BP groups (applicable only for chromosomal level)}
}
\description{
Interactive 3-D association landscape for many phenotypes
}
\examples{
phenos <- c("HDL", "LDL", "TRIGS", "TOTALCHOLESTROL")
## pass the dataframe from the processphegwas
y <- fastprocessphegwas(phenos)


3D landscape visualization of all the phenotypes across the base pair positions(above a threshold of -log10 (p) 6)
landscapefast(y,sliceval = 10,phenos =phenos)

3D landscape visualization of chromosome number 19 (above a threshold of -log10 (p) 10)
landscapefast(y,sliceval = 7.5,chromosome = 19,phenos =phenos)

3D landscape visualization of chromosome number 19, gene view active  (above a threshold of -log10 (p) 10)
landscape(y,sliceval = 7.5,chromosome = 19, geneview = TRUE,phenos =phenos)

3D visualization with LD block (for european population) passing externally, parameter to pass LD and also calculate the mutualLD block
landscapefast(y, sliceval = 30, chromosome = 19,calculateLD= TRUE,mutualLD = TRUE,phenos =phenos)
}
\author{
George Gittu
}
