% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PheGWAS2.R
\name{ldscmod}
\alias{ldscmod}
\title{Prepare the dataframe to pass to landscape function}
\usage{
ldscmod(phenos, filenames, N, hm3, ld)
}
\arguments{
\item{phenos}{Vector of names of dataframes that need to do iPheGWAS on.}

\item{dentogram}{to show structural differences}
}
\value{
A processed dataframe to pass to PheGWAS landscape function
}
\description{
Prepare the dataframe to pass to landscape function
}
\details{
Make sure there are no duplicate rsid's in any of the dataframe, If there aremake sure to resolve it before passing it to this function.
}
\examples{
\dontrun{
phenos <- c("HDL", "LDL", "TRIGS", "TOTALCHOLESTROL")
filennames <- c("HDL", "LDL", "TRIGS", "TOTALCHOLESTROL")
hm3 = "/Users/ggeorge/Desktop/ldscR/w_hm3.snplist"
ld = "/Users/ggeorge/Desktop/ldscR/MIX/eur_w_ld_chr"
N = c(53293,46350,51710,143677)
correlation.coeff <- ldscmod(phenos,filenames,N,hm3,ld)
}
}
\author{
George Gittu
}
