% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphegwasData.R
\docType{data}
\name{skeltonsnps}
\alias{skeltonsnps}
\title{Skeleton SNP's for ipheGWAS}
\format{
A dataframe with Skeleton SNP's (simulated using plink 1.9)
\itemize{
  \item rsid is the SNP
  \item A1 is allele1
  \item A1 is allele2
}
}
\usage{
data(skeltonsnps)
}
\value{
A data frame
}
\description{
Skeleton SNP's for ipheGWAS
}
\keyword{internal}
