% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphegwas.R
\name{iphegwas}
\alias{iphegwas}
\title{Prepare the dataframe to pass to landscape function}
\usage{
iphegwas(phenos, dentogram = FALSE, pathname = FALSE)
}
\arguments{
\item{phenos}{Vector of names of dataframes that need to do iPheGWAS on.}

\item{dentogram}{to show structural differences}

\item{pathname}{To read summarystats from a folder instead of passing it as phenos}
}
\value{
A processed dataframe to pass to PheGWAS landscape function/ dentogram if it's TRUE
}
\description{
Prepare the dataframe to pass to landscape function
}
\details{
Make sure there are no duplicate rsid's in any of the dataframe, If there aremake sure to resolve it before passing it to this function.
}
\examples{
\dontrun{
phenos <- c("HDL", "LDL", "TRIGS", "TOTALCHOLESTROL")
## This gives the order after applying iPheGWAS. We can pass this order to the landscape function
iphegwas(phenos)
## This gives the dentogram as output
iphegwas(phenos,dentogram = TRUE)
## Give folder path
pathname <- system.file("extdata", "samplesummary", package = "iphegwas")
iphegwas(pathname = pathname,dentogram = TRUE)
}
}
\author{
George Gittu
}
