% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphegwas.R
\name{ldscmod}
\alias{ldscmod}
\title{Function to use LDSC}
\usage{
ldscmod(pathname, ldscpath, dentogram = FALSE, plot = FALSE)
}
\arguments{
\item{pathname}{The folder that contains the summary stats to use for LDSC}

\item{ldscpath}{Path to the LDSC GitHub folder in your computer}

\item{dentogram}{If TRUE returns dentogram using LDSC method}

\item{plot}{If TRUE returns rg correlation plot using LDSC method}
}
\value{
Default LDSC ordering/ dentogram is TRUE/ rg plot if TRUE
}
\description{
Function to use LDSC
}
\details{
if used for LDSC ordering then summary stats file names should match with the phenos that you are passing to the landscape function.

You should have a ldsc conda environment that you use for running LDSC. Details here https://github.com/bulik/ldsc
}
\examples{
\dontrun{
## Giving absolute file path
pathname <- "/Users/ggeorg02/Desktop/databetasefiltered/files_ldsc/"
ldscpath <- "/Users/ggeorg02/Desktop/GitHub/ldsc"
# This gives the order after applying LDSC. We can pass this order to the landscape function
ldscmod(pathname,ldscpath)
ldscmod(pathname,ldscpath,dentogram = TRUE)
ldscmod(pathname,ldscpath,plot = TRUE)
}
}
\author{
George Gittu
}
